bank login:
LoginConroller registeres the user and creates an account in his name, SecurityFilter makes shure user that is not loged in is not allowed to enter secured locations, and user that is loged in does not log in or register second time. CurrentSessionProvider, is for identifying the current user using the system.