bank: morph ServletResponseWrite into ServletPageRenderer
A design of page rendering is improved by changing name and method
signature of the ServletResponseWriter.

The implementation now is responsible for loading of the page and
rendering it to the ServletResponse.

The duplication for loading of the pages which was in the init() method
of the servlets is now removed by this change.