bank: change misleading name of index page to registration page
The misleading name of Index page is now changed to Registration page
which is and the role of the page.

The app no longer serves "/" as expected, but with the integration with
security it will be handled as follow:

User tries to open "/" and if it's not authenticated he will be redirected
to the login page for login.

If he is authenticated then a proper home page should be rendered on the
screen.